requirements.txt
discord.py
scikit-video
opencv-python
